refactor(go): move configuration responsibilities (GO-SRP-1)
